+++ conda/meta.yaml
@@ -11,7 +11,7 @@ build:
   noarch: python
   preserve_egg_dir: True
   number: 1
-  script: python setup.py install 
+  script: python old_setup.py install
 
 requirements:
   build:
