+++ doc/api.rst
@@ -15,14 +15,6 @@ Reference guide
     :inherited-members:
     :show-inheritance:
 
-:mod:`openalea.astk.colormap` module
-===========================
-.. automodule:: openalea.astk.colormap
-    :members:
-    :undoc-members:
-    :inherited-members:
-    :show-inheritance:
-
 :mod:`openalea.astk.data_access` module
 ===========================
 .. automodule:: openalea.astk.data_access
@@ -39,22 +31,6 @@ Reference guide
     :inherited-members:
     :show-inheritance:
 
-:mod:`openalea.astk.plant_interface` module
-===========================
-.. automodule:: openalea.astk.plant_interface
-    :members:
-    :undoc-members:
-    :inherited-members:
-    :show-inheritance:
-
-:mod:`openalea.astk.plantgl_utils` module
-===========================
-.. automodule:: openalea.astk.plantgl_utils
-    :members:
-    :undoc-members:
-    :inherited-members:
-    :show-inheritance:
-
 :mod:`openalea.astk.sky_map` module
 ===========================
 .. automodule:: openalea.astk.sky_map
@@ -62,15 +38,6 @@ Reference guide
     :undoc-members:
     :inherited-members:
     :show-inheritance:
-
-:mod:`openalea.astk.sun_and_sky` module
-===========================
-.. automodule:: openalea.astk.sun_and_sky
-    :members:
-    :undoc-members:
-    :inherited-members:
-    :show-inheritance:
-
 
 :mod:`openalea.astk.TimeControl` module
 ===========================
@@ -88,33 +55,33 @@ Reference guide
     :inherited-members:
     :show-inheritance:
 
-:mod:`openalea.astk.meteorology.sky_irradiance` module
+:mod:`openalea.astk.sky_irradiance` module
 ===========================
-.. automodule:: openalea.astk.meteorology.sky_irradiance
+.. automodule:: openalea.astk.sky_irradiance
     :members:
     :undoc-members:
     :inherited-members:
     :show-inheritance:
 
-:mod:`openalea.astk.meteorology.sky_luminance` module
+:mod:`openalea.astk.sky_luminance` module
 ===========================
-.. automodule:: openalea.astk.meteorology.sky_luminance
+.. automodule:: openalea.astk.sky_luminance
     :members:
     :undoc-members:
     :inherited-members:
     :show-inheritance:
 
-:mod:`openalea.astk.meteorology.sun_position` module
+:mod:`openalea.astk.sun_position` module
 ===========================
-.. automodule:: openalea.astk.meteorology.sun_position
+.. automodule:: openalea.astk.sun_position
     :members:
     :undoc-members:
     :inherited-members:
     :show-inheritance:
 
-:mod:`openalea.astk.meteorology.sun_position_astk` module
+:mod:`openalea.astk.sun_position_astk` module
 ===========================
-.. automodule:: openalea.astk.meteorology.sun_position_astk
+.. automodule:: openalea.astk.sun_position_astk
     :members:
     :undoc-members:
     :inherited-members:
