+++ doc/conf.py
@@ -32,7 +32,7 @@ nbsphinx_allow_errors = True
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 autosummary_generate = True
-exclude_patterns = ["_build", "_templates"]
+exclude_patterns = ["build", "_build", "_templates"]
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 source_suffix = {
