+++ src/openalea/astk/__init__.py
@@ -1,7 +1,7 @@
 # {# pkglts, base
 
-from . import version
+from .version import __version__
 
-__version__ = version.__version__
+
 
 # #}
